Fail: real-time currency update

+++ app/src/main/java/kz/kd/hw_105/convertor/IFUpdateCurrencyList.kt
@@ -1,7 +1,5 @@
 package kz.kd.hw_105.convertor
 
-import android.widget.TextView
-
 interface IFUpdateCurrencyList {
-    fun updateCurrencyList(countryName: String, currencyAmount: Int)
+    fun updateCurrencyList(countryName: String, currencyAmount: Double)
 }
